Added Jawbone Distance measure

+++ UpPlatformSdk/src/main/java/com/jawbone/upplatformsdk/api/RestApiInterface.java
@@ -84,7 +84,7 @@ public interface RestApiInterface {
     @GET("/nudge/api/{version}/users/@me/moves")
     void getMoveEventsList(
             @Path(UpPlatformSdkConstants.API_VERSION) String version,
-            @QueryMap HashMap<String, Integer> hashMap,
+            @QueryMap HashMap<String, Long> hashMap,
             Callback<Object> response
     );
